+++ src/SwaggerProvider.DesignTime/v2/DefinitionCompiler.fs
@@ -29,8 +29,8 @@ type DefinitionPath =
             if ind = definitionPath.Length then
                 ind - 1
             elif
-                Char.IsLetterOrDigit definitionPath.[ind]
-                || definitionPath.[ind] = nsSeparator
+                Char.IsLetterOrDigit definitionPath[ind]
+                || definitionPath[ind] = nsSeparator
             then
                 getCharInTypeName(ind + 1)
             else
@@ -76,10 +76,10 @@ and NamespaceAbstraction(name: string) =
         | _, value -> failwithf $"Cannot %s{opName} '%s{tyName}' because the slot is used by %A{value}"
 
     /// Namespace name
-    member __.Name = name
+    member _.Name = name
 
     /// Generate unique name and reserve it for the type
-    member __.ReserveUniqueName namePref nameSuffix = // TODO: Strange signature - think more
+    member _.ReserveUniqueName namePref nameSuffix = // TODO: Strange signature - think more
         let rec findUniq prefix i =
             let newName = sprintf "%s%s" prefix (if i = 0 then "" else i.ToString())
 
@@ -93,34 +93,34 @@ and NamespaceAbstraction(name: string) =
         newName
 
     /// Release previously reserved name
-    member __.ReleaseNameReservation tyName =
+    member _.ReleaseNameReservation tyName =
         updateReservation "release the name" tyName (fun () -> providedTys.Remove(tyName) |> ignore)
 
     /// Mark type name as named alias for basic type
-    member __.MarkTypeAsNameAlias tyName =
-        updateReservation "mark as Alias type" tyName (fun () -> providedTys.[tyName] <- NameAlias)
+    member _.MarkTypeAsNameAlias tyName =
+        updateReservation "mark as Alias type" tyName (fun () -> providedTys[tyName] <- NameAlias)
 
     /// Associate ProvidedType with reserved type name
-    member __.RegisterType(tyName, ty) =
+    member _.RegisterType(tyName, ty) =
         match providedTys.TryGetValue tyName with
-        | true, Reservation -> providedTys.[tyName] <- ProvidedType ty
-        | true, Namespace ns -> providedTys.[tyName] <- NestedType(ty, ns)
+        | true, Reservation -> providedTys[tyName] <- ProvidedType ty
+        | true, Namespace ns -> providedTys[tyName] <- NestedType(ty, ns)
         | false, _ -> failwithf $"Cannot register the type '%s{tyName}' because name was not reserved"
         | _, value -> failwithf $"Cannot register the type '%s{tyName}' because the slot is used by %A{value}"
 
     /// Get or create sub-namespace
-    member __.GetOrCreateNamespace name =
+    member _.GetOrCreateNamespace name =
         match providedTys.TryGetValue name with
         | true, Namespace ns -> ns
         | true, NestedType(_, ns) -> ns
         | true, ProvidedType ty ->
             let ns = NamespaceAbstraction(name)
-            providedTys.[name] <- NestedType(ty, ns)
+            providedTys[name] <- NestedType(ty, ns)
             ns
         | false, _
         | true, Reservation ->
             let ns = NamespaceAbstraction(name)
-            providedTys.[name] <- Namespace ns
+            providedTys[name] <- Namespace ns
             ns
         | true, value -> failwithf $"Name collision, cannot create namespace '%s{name}' because it used by '%A{value}'"
 
@@ -133,7 +133,7 @@ and NamespaceAbstraction(name: string) =
             ns.Resolve { dPath with Namespace = tail }
 
     /// Create Provided representation of Namespace
-    member __.GetProvidedTypes() =
+    member _.GetProvidedTypes() =
         List.ofSeq providedTys
         |> List.choose(fun kv ->
             match kv.Value with
@@ -168,7 +168,7 @@ type DefinitionCompiler(schema: SwaggerObject, provideNullable) as this =
         let propertyName = scope.MakeUnique <| nicePascalName propName
 
         let providedField =
-            let fieldName = $"_%c{Char.ToLower propertyName.[0]}%s{propertyName.Substring(1)}"
+            let fieldName = $"_%c{Char.ToLower propertyName[0]}%s{propertyName.Substring(1)}"
 
             ProvidedField(fieldName, ty)
 
@@ -239,7 +239,7 @@ type DefinitionCompiler(schema: SwaggerObject, provideNullable) as this =
                         let pTy =
                             compileSchemaObject ns (ns.ReserveUniqueName tyName (nicePascalName p.Name)) p.Type p.IsRequired ns.RegisterType
 
-                        let (pField, pProp) = generateProperty p.Name pTy
+                        let pField, pProp = generateProperty p.Name pTy
 
                         if not <| String.IsNullOrWhiteSpace p.Description then
                             pProp.AddXmlDoc p.Description
@@ -279,7 +279,7 @@ type DefinitionCompiler(schema: SwaggerObject, provideNullable) as this =
                     ctorParams,
                     invokeCode =
                         fun args ->
-                            let (this, args) =
+                            let this, args =
                                 match args with
                                 | x :: xs -> (x, xs)
                                 | _ -> failwith "Wrong constructor arguments"
@@ -287,7 +287,7 @@ type DefinitionCompiler(schema: SwaggerObject, provideNullable) as this =
                             List.zip args fields
                             |> List.map(fun (arg, f) -> Expr.FieldSetUnchecked(this, f, arg))
                             |> List.rev
-                            |> List.fold (fun a b -> Expr.Sequential(a, b)) (<@@ () @@>)
+                            |> List.fold (fun a b -> Expr.Sequential(a, b)) <@@ () @@>
                 )
 
                 // Override `.ToString()`
@@ -299,9 +299,9 @@ type DefinitionCompiler(schema: SwaggerObject, provideNullable) as this =
                         isStatic = false,
                         invokeCode =
                             fun args ->
-                                let this = args.[0]
+                                let this = args[0]
 
-                                let (pNames, pValues) =
+                                let pNames, pValues =
                                     Array.ofList members
                                     |> Array.map(fun (pField, pProp) ->
                                         let pValObj = Expr.FieldGet(this, pField)
@@ -329,7 +329,7 @@ type DefinitionCompiler(schema: SwaggerObject, provideNullable) as this =
 
                                     let strs =
                                         values
-                                        |> Array.mapi(fun i v -> String.Format("{0}={1}", pNames.[i], formatValue v))
+                                        |> Array.mapi(fun i v -> String.Format("{0}={1}", pNames[i], formatValue v))
 
                                     String.Format("{{{0}}}", String.Join("; ", strs))
                                 @@>
@@ -337,7 +337,7 @@ type DefinitionCompiler(schema: SwaggerObject, provideNullable) as this =
 
                 toStr.SetMethodAttrs(MethodAttributes.Public ||| MethodAttributes.Virtual)
 
-                let objToStr = (typeof<obj>).GetMethod("ToString", [||])
+                let objToStr = typeof<obj>.GetMethod ("ToString", [||])
                 ty.DefineMethodOverride(toStr, objToStr)
                 ty.AddMember <| toStr
 
@@ -362,10 +362,10 @@ type DefinitionCompiler(schema: SwaggerObject, provideNullable) as this =
                 | String -> typeof<string>
                 | Date
                 | DateTime -> typeof<DateTime>
-                | File -> typeof<byte>.MakeArrayType (1)
+                | File -> typeof<byte>.MakeArrayType 1
                 | Enum(_, "string") -> typeof<string>
                 | Enum(_, "boolean") -> typeof<bool>
-                | Enum(_, _) -> typeof<int32>
+                | Enum _ -> typeof<int32>
                 | Array eTy ->
                     (compileSchemaObject ns (ns.ReserveUniqueName tyName "Item") eTy true ns.RegisterType)
                         .MakeArrayType(1)
@@ -399,14 +399,14 @@ type DefinitionCompiler(schema: SwaggerObject, provideNullable) as this =
         |> Seq.iter(fun (name, _) -> compileDefinition name |> ignore)
 
     /// Namespace that represent provided type space
-    member __.Namespace = nsRoot
+    member _.Namespace = nsRoot
 
     /// Method that allow OperationCompiler to resolve object reference, compile basic and anonymous types.
-    member __.CompileTy opName tyUseSuffix ty required =
+    member _.CompileTy opName tyUseSuffix ty required =
         compileSchemaObject nsOps (nsOps.ReserveUniqueName opName tyUseSuffix) ty required nsOps.RegisterType
 
     /// Default value for optional parameters
-    member __.GetDefaultValue _ =
+    member _.GetDefaultValue _ =
         // This method is only used for not required types
         // Reference types, Option<T> and Nullable<T>
         null