+++ src/SwaggerProvider.Runtime/RuntimeHelpers.fs
@@ -2,7 +2,6 @@ namespace Swagger.Internal
 
 open System
 open System.Net.Http
-open System.Text.Json
 open System.Text.Json.Serialization
 open System.Threading.Tasks
 
@@ -12,6 +11,13 @@ module MediaTypes =
 
     [<Literal>]
     let ApplicationOctetStream = "application/octet-stream"
+
+    [<Literal>]
+    let ApplicationFormUrlEncoded = "application/x-www-form-urlencoded"
+
+    [<Literal>]
+    let MultipartFormData = "multipart/form-data"
+
 
 type AsyncExtensions() =
     static member cast<'t> asyncOp = async {
@@ -107,17 +113,22 @@ module RuntimeHelpers =
 
     let getPropertyNameAttribute name =
         { new Reflection.CustomAttributeData() with
-            member __.Constructor =
-                typeof<JsonPropertyNameAttribute>.GetConstructor ([| typeof<string> |])
+            member _.Constructor =
+                typeof<JsonPropertyNameAttribute>.GetConstructor [| typeof<string> |]
 
-            member __.ConstructorArguments =
+            member _.ConstructorArguments =
                 [| Reflection.CustomAttributeTypedArgument(typeof<string>, name) |] :> Collections.Generic.IList<_>
 
-            member __.NamedArguments = [||] :> Collections.Generic.IList<_>
+            member _.NamedArguments = [||] :> Collections.Generic.IList<_>
         }
 
     let toStringContent(valueStr: string) =
         new StringContent(valueStr, Text.Encoding.UTF8, "application/json")
+
+    let toStreamContent(boxedStream: obj) =
+        match boxedStream with
+        | :? IO.Stream as stream -> new StreamContent(stream)
+        | _ -> failwith $"Unexpected parameter type {boxedStream.GetType().Name} instead of IO.Stream"
 
     let getPropertyValues(object: obj) =
         if isNull object then
@@ -146,7 +157,7 @@ module RuntimeHelpers =
             let filename = Guid.NewGuid().ToString() // asp.net core cannot deserialize IFormFile otherwise
             cnt.Add(new StreamContent(stream), name, filename)
 
-        for (name, value) in keyValues do
+        for name, value in keyValues do
             match value with
             | null -> ()
             | :? IO.Stream as stream -> addFileStream name stream
@@ -185,7 +196,7 @@ module RuntimeHelpers =
             let builder = UriBuilder(combineUrl fakeHost address)
             let query = System.Web.HttpUtility.ParseQueryString(builder.Query)
 
-            for (name, value) in queryParams do
+            for name, value in queryParams do
                 if not <| isNull value then
                     query.Add(name, value)
 
@@ -207,11 +218,11 @@ module RuntimeHelpers =
                     raise <| Exception(errMsg))
 
     let asyncCast runtimeTy (asyncOp: Async<obj>) =
-        let castFn = typeof<AsyncExtensions>.GetMethod ("cast")
+        let castFn = typeof<AsyncExtensions>.GetMethod "cast"
 
         castFn.MakeGenericMethod([| runtimeTy |]).Invoke(null, [| asyncOp |])
 
     let taskCast runtimeTy (task: Task<obj>) =
-        let castFn = typeof<TaskExtensions>.GetMethod ("cast")
+        let castFn = typeof<TaskExtensions>.GetMethod "cast"
 
         castFn.MakeGenericMethod([| runtimeTy |]).Invoke(null, [| task |])