+++ tests/Swashbuckle.WebApi.Server/Controllers/FileController.fs
@@ -1,13 +1,33 @@
 namespace Swashbuckle.WebApi.Server.Controllers
 
 open System
+open System.IO
 open Microsoft.AspNetCore.Mvc
 open Microsoft.AspNetCore.Http
+open Microsoft.OpenApi.Models
 open Swagger.Internal
+open Swashbuckle.AspNetCore.SwaggerGen
 
 type FormWithFile() =
     member val Name: string = "" with get, set
     member val File: IFormFile = null with get, set
+
+// https://stackoverflow.com/questions/41141137/how-can-i-tell-swashbuckle-that-the-body-content-is-required
+type BinaryContentAttribute() =
+    inherit Attribute()
+
+type BinaryContentFilter() =
+    interface IOperationFilter with
+        member _.Apply(op, ctx) =
+            let att = ctx.MethodInfo.GetCustomAttributes(typeof<BinaryContentAttribute>, false)
+
+            if att.Length > 0 then
+                op.RequestBody <- OpenApiRequestBody(Required = true)
+
+                op.RequestBody.Content.Add(
+                    MediaTypes.ApplicationOctetStream,
+                    OpenApiMediaType(Schema = OpenApiSchema(Type = "string", Format = "binary"))
+                )
 
 [<Route("api/[controller]")>]
 [<ApiController>]
@@ -17,8 +37,15 @@ type ReturnFileController() =
     [<HttpGet; Produces(MediaTypes.ApplicationOctetStream, Type = typeof<FileResult>)>]
     member this.Get() =
         let bytes = System.Text.Encoding.UTF8.GetBytes("I am totally a file's\ncontent")
-        let stream = new System.IO.MemoryStream(bytes)
+        let stream = new MemoryStream(bytes)
         this.File(stream, MediaTypes.ApplicationOctetStream, "hello.txt") :> FileResult
+
+    [<HttpPost("stream"); BinaryContent>]
+    member this.GetFileLength() = task {
+        use reader = new StreamReader(this.Request.Body)
+        let! content = reader.ReadToEndAsync()
+        return content.Length
+    }
 
     [<HttpPost("single"); Produces(MediaTypes.ApplicationOctetStream, Type = typeof<FileResult>)>]
     member this.PostFile(file: IFormFile) : FileResult =